refactor(directoryTable): Splitting row into components for loading state
